+++ libpostal/cppbuild.sh
@@ -21,6 +21,9 @@ if [[ "${ACLOCAL_PATH:-}" == C:\\msys64\\* ]]; then
     export ACLOCAL_PATH=/mingw64/share/aclocal:/usr/share/aclocal
 fi
 
+# Work around build issues on Windows
+sedinplace '/_rand48_/d' src/klib/drand48.h
+
 case $PLATFORM in
     linux-x86_64)
         ./bootstrap.sh