+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Update presets for Visual Studio 2019 on Windows
  * Add presets for TVM 0.7.0
  * Include `free()` in presets for FTTW as required by `fftw_export_wisdom_to_string()` ([issue bytedeco/javacpp#429](https://github.com/bytedeco/javacpp/issues/429))
  * Include all missing header files from the `opencv_ximgproc` module ([issue #958](https://github.com/bytedeco/javacpp-presets/issues/958))